Fix path traversal, dead code, and token duplication in proxy/web
1. Path traversal: replace naive string check with resolve() + prefix
   guard in both proxy.ts and web.ts, handles encoded sequences
2. Remove unused staticFiles map from proxy.ts
3. Extract getProxyToken() to global-config.ts shared helper —
   daemon and server now both check KERN_WEB_TOKEN fallback
