+++ src/global-config.ts
@@ -1,7 +1,8 @@
-import { readFile, writeFile, mkdir, unlink } from "fs/promises";
+import { readFile, writeFile, mkdir, unlink, appendFile } from "fs/promises";
 import { join } from "path";
 import { existsSync, readFileSync } from "fs";
 import { homedir } from "os";
+import { randomBytes } from "crypto";
 import { log } from "./log.js";
 
 export interface GlobalConfig {
@@ -82,4 +83,21 @@ async function migrateLegacyAgents(): Promise<void> {
   } catch (err) {
     log.warn("config", `agents.json migration failed: ${err}`);
   }
+}
+
+const ENV_FILE = join(homedir(), ".kern", ".env");
+
+/** Load or auto-generate the proxy auth token from ~/.kern/.env */
+export async function getProxyToken(): Promise<string> {
+  if (existsSync(ENV_FILE)) {
+    const content = await readFile(ENV_FILE, "utf-8");
+    // Check new name first, fall back to legacy KERN_WEB_TOKEN
+    const match = content.match(/^KERN_PROXY_TOKEN=(.+)$/m)
+      || content.match(/^KERN_WEB_TOKEN=(.+)$/m);
+    if (match) return match[1].trim();
+  }
+  const token = randomBytes(16).toString("hex");
+  await appendFile(ENV_FILE, `${existsSync(ENV_FILE) ? "\n" : ""}KERN_PROXY_TOKEN=${token}\n`);
+  log("proxy", `generated proxy token: ${token.slice(0, 8)}...`);
+  return token;
 }
